refactor: replaced schema by measure validation
